Example of SERVICE statement
It is a good idea to include an example of LOD's federated power queries. Because the BM does not link out to rich sources like this yet, I'm using an example from Europeana, with the caveat that the lesson won't be going into the Europeana or DBpedia data model in depth. Hopefully this sample query, and the sequence diagram, can help readers understand how a SELECT statement can work in the real world.