lesson-slug.html now takes a variable
This change allows a lesson page object to be passed to the slug
generator from the lesson index page. If there is no provided
lesson object - which will happen when there is just a simple figure
include in a standard lesson, then lesson-slug defaults to asking
for the page.url. If one is provided, though, it will use that specific
page. This is useful when writing liquid that iterates through lessons,
like the lesson index does.