Update miditime import
The miditime import in the example didn't work for me, I had to switch it to this version.  I used `pip install` and got `miditime==1.1.3`.    (The project documentation shows the old import, but that doesn't work.)

I thought it might be worth updating, because if people aren't familiar with Python they may not be able to figure it out.   You might want to add something about the version of miditime or document the older syntax also, but I don't know where that should go.