Fixed GitHub Icon (#845)
GitHub icon was rendering as a block because the naming in Font Awesome 5 was slightly different. https://fontawesome.com/icons/github?style=brands

There is not an icon associated with `fa fa-github` in Font Awesome 5 but there is an icon associated with`fab fa-github`. It appears the former was supported in Font Awesome 4 but Programming Historian is now running on Font Awesome 5.