Use same prefixes for RDF/XML as before
In the XML document, the lccn URI mixed up the HTTP scheme with the
INFO scheme.  If the INFO scheme really should be used here (although
it is not explained), it should probably read `info:lccn/2010025398`
(without the prefix `http://`).
The `viaf` URI missed `viaf` in its paths, to mirror earlier URIs in
turtle examples.
Also, the prefix `lccn` was missing twice.