Create translator-guidelines (#1260)
Closes #1199 

* Create translator-guidelines

* Update translator-guidelines

Adapted part of the author-guidelines to translator-guidelines.

* Update translator-guidelines

precised some of the tasks a translation involves

* Update and rename translator-guidelines to translator-guidelines.md

- Needs adjustment of link to raw text: by inspecting [the raw text on GitHub](en/translator-guidelines)

As you will note, most of the text resembles the original author guidelines, as the markdown and submission process are practically the same. This is to avoid having to switch between pages.

* Update translator-guidelines.md

Added a note on linguistic variations.

* Update translator-guidelines.md

Edited two links that were causing issues with Travis.