Create Doacoes.md
I create the page "Individual Supporters" in Portuguese. The link to the original page is https://programminghistorian.org/en/individual
It is part of the 3 new pages that will be created in Portuguese about fundraising.

By markdown editor, text formatting was ok. When pasting the text here on Github it appeared different. If necessary, I make the adjustment.