Update mapping-with-python-leaflet.md
lines 159-165: removing third column (labelled Google V3) from table
para. 167: adding the sentences "GoogleV3 is a geocoder compatible with geopy, it is a reliable geolocator choice because of their large geographic data coverage. However, since July 2018 an API key is required, and you need to enable billing in Google Cloud to use it."